# backend/src/docs/swagger.yaml
openapi: 3.0.3
info:
  title: Sea Hawk Courier API
  version: 1.0.0
  description: Core operational and client portal endpoints for Sea Hawk Courier.
servers:
  - url: http://localhost:3001
    description: Local
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
paths:
...
  /api/invoices:
    get:
      tags: [Invoices]
      summary: List invoices
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Invoice list
    post:
      tags: [Invoices]
      summary: Create invoice
      security:
        - bearerAuth: []
      responses:
        '201':
          description: Invoice created
  /api/invoices/{id}/pdf:
    get:
      tags: [Invoices]
      summary: Download invoice PDF
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: PDF file
...